Revert "Consolidate front‑end assets"
