+++ conViver.Web/pages/reservas.html
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang="pt-BR">
+<head>
+    <meta charset="UTF-8">
+    <title>conViver - Reservas</title>
+    <link rel="stylesheet" href="../css/components.css">
+    <link rel="stylesheet" href="../css/styles.css">
+</head>
+<body>
+    <header class="cv-header">
+        <h1>conViver</h1>
+    </header>
+    <nav class="cv-nav">
+        <a href="../index.html" class="cv-nav__link">Dashboard</a>
+        <a href="financeiro.html" class="cv-nav__link">Financeiro</a>
+        <a href="reservas.html" class="cv-nav__link">Reservas</a>
+        <a href="portaria.html" class="cv-nav__link">Portaria</a>
+        <a href="comunicacao.html" class="cv-nav__link">Comunicação</a>
+    </nav>
+    <main class="cv-container">
+        <section class="reservas__calendar js-calendario">
+            <h2 class="reservas__title">Agenda das Áreas Comuns</h2>
+            <div class="cv-card reservas__day">Dia 1</div>
+            <div class="cv-card reservas__day">Dia 2</div>
+        </section>
+        <section class="reservas__solicitar">
+            <h2 class="reservas__title">Solicitar Reserva</h2>
+            <form class="reservas__form js-reserva-form">
+                <label class="reservas__label">
+                    Data:
+                    <input type="date" class="reservas__input js-reserva-data" />
+                </label>
+                <button class="reservas__button" type="submit">Enviar</button>
+            </form>
+        </section>
+    </main>
+    <script type="module" src="../js/reservas.js"></script>
+</body>
+</html>
