+++ conViver.Web/wwwroot/js/auth.js
@@ -1,12 +1,1 @@
-import { getToken } from './apiClient.js';
-
-export function requireAuth() {
-    if (!getToken()) {
-        window.location.href = 'login.html';
-    }
-}
-
-export function logout() {
-    localStorage.removeItem('cv_token');
-    window.location.href = 'login.html';
-}
+console.log('auth loaded');
