+++ conViver.Web/wwwroot/pages/financeiro.html
@@ -2,96 +2,61 @@
 <html lang="pt-BR">
 <head>
     <meta charset="UTF-8">
-    <title>conViver - Gestão Financeira</title>
+    <title>conViver - Financeiro</title>
     <link rel="stylesheet" href="../css/components.css">
     <link rel="stylesheet" href="../css/styles.css">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
 </head>
-<body class="financeiro-page">
+<body>
     <header class="cv-header">
-        <h1 class="cv-header__title">conViver</h1>
-        <button id="logoutButton" class="cv-button cv-button--secondary cv-header__logout">Sair</button>
+        <h1>conViver</h1>
     </header>
     <nav class="cv-nav">
-        <a href="dashboard.html" class="cv-nav__link">Dashboard</a>
-        <a href="financeiro.html" class="cv-nav__link cv-nav__link--active">Financeiro</a>
+        <a href="../index.html" class="cv-nav__link">Dashboard</a>
+        <a href="financeiro.html" class="cv-nav__link">Financeiro</a>
         <a href="reservas.html" class="cv-nav__link">Reservas</a>
         <a href="portaria.html" class="cv-nav__link">Portaria</a>
         <a href="comunicacao.html" class="cv-nav__link">Comunicação</a>
     </nav>
-    <main class="cv-container fin-cobrancas-main">
-        <h1 class="fin-cobrancas__title">Emissão e Gestão de Cobranças</h1>
-
-        <section class="db-section fin-cobrancas__dashboard">
-            <h2 class="db-section__title">Resumo Financeiro</h2>
-            <div class="db-metrics-grid">
-                <div class="cv-card fin-summary-card">
-                    <h4 class="fin-summary-card__title">Inadimplência</h4>
-                    <p class="fin-summary-card__value js-summary-inadimplencia">--%</p>
-                </div>
-                <div class="cv-card fin-summary-card">
-                    <h4 class="fin-summary-card__title">Pix no Mês</h4>
-                    <p class="fin-summary-card__value js-summary-pix">R$ --</p>
-                </div>
-                <div class="cv-card fin-summary-card">
-                    <h4 class="fin-summary-card__title">Boletos Pendentes</h4>
-                    <p class="fin-summary-card__value js-summary-pendentes">--</p>
-                </div>
-            </div>
+    <main class="cv-container">
+        <section class="financial__section">
+            <h2 class="financial__title">Emissão e Gestão de Cobranças</h2>
+            <table class="cv-table financial__charges-table">
+                <thead>
+                    <tr>
+                        <th>Unidade</th>
+                        <th>Valor</th>
+                        <th>Status</th>
+                    </tr>
+                </thead>
+                <tbody class="js-boletos">
+                    <tr>
+                        <td>101</td>
+                        <td>R$ --</td>
+                        <td>Pendente</td>
+                    </tr>
+                </tbody>
+            </table>
         </section>
-
-        <section class="db-section fin-cobrancas__actions">
-            <h2 class="db-section__title visually-hidden">Ações de Cobrança</h2>
-            <div class="fin-actions__group">
-                <label for="filtroStatusCobranca">Status:</label>
-                <select id="filtroStatusCobranca" class="cv-input js-filtro-status-cobranca">
-                    <option value="">Todos</option>
-                    <option value="Pendente">Pendente</option>
-                    <option value="Pago">Pago</option>
-                    <option value="Atrasado">Atrasado</option>
-                    <option value="Cancelado">Cancelado</option>
-                </select>
-            </div>
-            <div class="fin-actions__group">
-                <button id="btnNovaCobranca" class="cv-button cv-button--primary js-btn-nova-cobranca">Emitir Nova Cobrança</button>
-                <button id="btnGerarLote" class="cv-button js-btn-gerar-lote">Gerar Lote Mensal</button>
-            </div>
+        <section class="financial__section">
+            <h2 class="financial__title">Controle de Despesas</h2>
+            <table class="cv-table financial__expenses-table">
+                <thead>
+                    <tr>
+                        <th>Descrição</th>
+                        <th>Vencimento</th>
+                        <th>Valor</th>
+                    </tr>
+                </thead>
+                <tbody class="js-despesas">
+                    <tr>
+                        <td>...</td>
+                        <td>...</td>
+                        <td>...</td>
+                    </tr>
+                </tbody>
+            </table>
         </section>
-
-        <section class="db-section fin-cobrancas__table-section">
-            <h2 class="db-section__title">Lista de Cobranças</h2>
-            <div class="table-responsive">
-                <table class="cv-table fin-cobrancas__table">
-                    <thead>
-                        <tr>
-                            <th>Unidade</th>
-                            <th>Sacado</th>
-                            <th>Valor</th>
-                            <th>Vencimento</th>
-                            <th>Status</th>
-                            <th>Ações</th>
-                        </tr>
-                    </thead>
-                    <tbody class="js-lista-cobrancas">
-                        <tr><td colspan="6">Carregando cobranças...</td></tr>
-                    </tbody>
-                </table>
-            </div>
-        </section>
-
-        <!-- Modal (inicialmente oculto) -->
-        <div id="modalCobranca" class="cv-modal js-modal-cobranca" style="display: none;">
-            <div class="cv-modal__content">
-                <span class="cv-modal__close js-modal-close-cobranca">&times;</span>
-                <h3 id="modalCobrancaTitle" class="cv-modal__title">Nova Cobrança</h3>
-                <div id="modalCobrancaBody" class="cv-modal__body">
-                    <!-- Formulário ou detalhes aqui -->
-                    <p>Conteúdo do modal...</p>
-                </div>
-            </div>
-        </div>
     </main>
     <script type="module" src="../js/financeiro.js"></script>
-    <!-- <script type="module" src="../js/main.js"></script> --> <!-- main.js likely imported by financeiro.js if needed -->
 </body>
 </html>
